+++ src/selmr.rs
@@ -1,11 +1,9 @@
 use crate::text_structs::{Context, ContextCounter, ContextMap, Phrase, PhraseCounter, PhraseMap};
 use crate::tokenizer;
-use core::cmp::min;
+use core::cmp::{max, min};
 use counter::Counter;
 use indexmap::IndexMap;
 use log::info;
-use pyo3::exceptions::PyTypeError;
-use pyo3::prelude::*;
 use regex::Regex;
 use serde::{ser::SerializeStruct, Serialize, Serializer};
 use std::collections::{HashMap, HashSet};
@@ -13,75 +11,102 @@ use std::io::prelude::*;
 use std::{fs, io::Write, path};
 use zip::{write::FileOptions, ZipArchive, ZipWriter};
 
-#[pymodule]
-fn pyselmr(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
-    pyo3_log::init();
-    m.add_class::<SELMR>()?;
-    Ok(())
-}
-
 /// Struct containing all parameters of a SELMR data structure
 #[derive(Deserialize, Serialize, Debug)]
 pub struct Params {
-    min_phrase_len: usize,
-    max_phrase_len: usize,
-    min_context_len: usize,
-    max_context_len: usize,
-    min_phrase_keys: usize,
-    min_context_keys: usize,
-    language: String,
+    /// minimum number of words in a phrase
+    pub min_phrase_len: usize,
+    /// maximum number of words in a phrase
+    pub max_phrase_len: usize,
+    /// minimum number of words in the left part of a context
+    pub min_left_context_len: usize,
+    /// maximum number of words in the left part of a context
+    pub max_left_context_len: usize,
+    /// minimum number of words in the right part of a context
+    pub min_right_context_len: usize,
+    /// maximum number of words in the right part of a context
+    pub max_right_context_len: usize,
+    /// minimum number of contexts in which a phrase must fit
+    pub min_phrase_keys: usize,
+    /// minimum number of phrases that must fit into a context
+    pub min_context_keys: usize,
+    /// the language of the phrases and contexts
+    pub language: String,
 }
 
 /// The SELMR data structure
 #[derive(Deserialize)]
-#[pyclass]
 pub struct SELMR {
-    phrases: PhraseMap,
+    /// the phrases to contexts map
+    pub phrases: PhraseMap,
+    /// the contexts to phrases map
     #[serde(skip)]
-    contexts: ContextMap,
-    params: Params,
+    pub contexts: ContextMap,
+    /// the parameters of the SELMR struct
+    pub params: Params,
+}
+
+impl Default for SELMR {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+#[inline]
+fn count_index<T: std::cmp::Eq + std::hash::Hash>(a: &HashSet<&T>, b: &HashSet<&T>) -> f32 {
+    a.intersection(b).collect::<Vec<_>>().len() as f32
+}
+
+#[inline]
+fn jaccard_index<T: std::cmp::Eq + std::hash::Hash>(a: &HashSet<&T>, b: &HashSet<&T>) -> f32 {
+    a.intersection(b).collect::<Vec<_>>().len() as f32 / max(a.len(), b.len()) as f32
+}
+
+#[inline]
+fn weighted_jaccard_index<T: std::cmp::Eq + std::hash::Hash>(
+    a: &HashMap<&T, &usize>,
+    b: &HashMap<&T, &usize>,
+) -> f32 {
+    let mut num: f32 = 0.0;
+    let mut denom: f32 = 0.0;
+    for context in a
+        .keys()
+        .collect::<HashSet<_>>()
+        .union(&b.keys().collect())
+    {
+        num += *min(*(a.get(*context).unwrap_or(&&0)), *(b.get(*context).unwrap_or(&&0))) as f32;
+        denom += *max(*(a.get(*context).unwrap_or(&&0)), *(b.get(*context).unwrap_or(&&0))) as f32;
+    }
+    if denom != 0.0 {
+        num / denom
+    } else {
+        0.0
+    }
 }
 
 // Core SELMR methods
-#[pymethods]
 impl SELMR {
     /// Initializes an empty SELMR data structure
     ///
     /// # Example
     ///
     /// ```
-    /// let s = selmr::selmr::SELMR::new(1, 3, 1, 3, 1, 1, "en");
+    /// let s = selmr::selmr::SELMR::new();
     /// ```
-    #[new]
-    #[pyo3(signature = (
-        min_phrase_len=1,
-        max_phrase_len=3,
-        min_context_len=1,
-        max_context_len=3,
-        min_phrase_keys=2,
-        min_context_keys=2,
-        language="en")
-    )]
-    pub fn new(
-        min_phrase_len: usize,
-        max_phrase_len: usize,
-        min_context_len: usize,
-        max_context_len: usize,
-        min_phrase_keys: usize,
-        min_context_keys: usize,
-        language: &str,
-    ) -> Self {
+    pub fn new() -> Self {
         SELMR {
             phrases: PhraseMap::new(),
             contexts: ContextMap::new(),
             params: Params {
-                min_phrase_len,
-                max_phrase_len,
-                min_context_len,
-                max_context_len,
-                min_phrase_keys,
-                min_context_keys,
-                language: language.to_string(),
+                min_phrase_len: 1,
+                max_phrase_len: 3,
+                min_left_context_len: 1,
+                max_left_context_len: 3,
+                min_right_context_len: 1,
+                max_right_context_len: 3,
+                min_phrase_keys: 5,
+                min_context_keys: 5,
+                language: "en".to_string(),
             },
         }
     }
@@ -91,10 +116,10 @@ impl SELMR {
     /// # Example
     ///
     /// ```
-    /// let s = selmr::selmr::SELMR::new(1, 3, 1, 3, 1, 1, "en");
+    /// let s = selmr::selmr::SELMR::new();
     /// s.write("test.json", "json").expect("REASON");
     /// ```
-    pub fn write(&self, file: &str, format: &str) -> Result<(), PyErr> {
+    pub fn write(&self, file: &str, format: &str) -> Result<(), std::io::Error> {
         let mut f = fs::File::create(file)?;
         if format == "json" {
             let s = serde_json::to_string_pretty(&self).unwrap();
@@ -106,23 +131,11 @@ impl SELMR {
                 Ok(()) => match zip.write_all(s.as_bytes()) {
                     Ok(()) => match zip.finish() {
                         Ok(_) => return Ok(()),
-                        Err(_) => {
-                            return Err(PyErr::new::<PyTypeError, _>(
-                                "error closing finishing writing file in zip file",
-                            ))
-                        }
+                        Err(e) => return Err(e.into()),
                     },
-                    Err(_) => {
-                        return Err(PyErr::new::<PyTypeError, _>(
-                            "error writing to file in zip file",
-                        ))
-                    }
+                    Err(e) => return Err(e),
                 },
-                Err(_) => {
-                    return Err(PyErr::new::<PyTypeError, _>(
-                        "data.json not found in zip file",
-                    ))
-                }
+                Err(e) => return Err(e.into()),
             }
         }
         Ok(())
@@ -133,10 +146,10 @@ impl SELMR {
     /// # Example
     ///
     /// ```
-    /// let mut s = selmr::selmr::SELMR::new(1, 3, 1, 3, 1, 1, "en");
+    /// let mut s = selmr::selmr::SELMR::new();
     /// s.read("test.json", "json").expect("REASON");
     /// ```
-    pub fn read(&mut self, file: &str, format: &str) -> Result<(), PyErr> {
+    pub fn read(&mut self, file: &str, format: &str) -> Result<(), std::io::Error> {
         let mut s: Option<SELMR> = None;
         if format == "json" {
             let data = fs::read_to_string(file);
@@ -147,10 +160,10 @@ impl SELMR {
                         Ok(from_str) => {
                             s = from_str;
                         }
-                        Err(_) => return Err(PyErr::new::<PyTypeError, _>("Unable to parse")),
+                        Err(e) => return Err(e.into()),
                     }
                 }
-                Err(e) => return Err(PyErr::new::<PyTypeError, _>(e)),
+                Err(e) => return Err(e),
             }
         } else if format == "zip" {
             let path = path::Path::new(file);
@@ -166,42 +179,61 @@ impl SELMR {
                             s = from_str;
                         }
                         Err(e) => {
-                            return Err(PyErr::new::<PyTypeError, _>(e.to_string()));
+                            return Err(e.into());
                         }
                     }
                 }
-                Err(..) => {
-                    return Err(PyErr::new::<PyTypeError, _>(
-                        "data.json not found in zipfile.",
-                    ));
+                Err(e) => {
+                    return Err(e.into());
                 }
             };
         }
-        // the self.contexts is derived from self.phrases, and must be ordered
         if let Some(selmr) = s {
-            *self = selmr;
-            let mut unordered_contexts = HashMap::<Context, Counter<Phrase>>::new();
-            for (p, contexts) in self.phrases.map.iter() {
-                for (c, &n) in contexts.map.iter() {
-                    unordered_contexts
-                        .entry(c.clone())
-                        .or_default()
-                        .insert(p.clone(), n);
-                }
-            }
-            self.contexts = ContextMap::new();
-            for (context, context_phrases) in unordered_contexts.iter() {
-                let ordered_context_phrases: IndexMap<Phrase, usize> =
-                    context_phrases.most_common_ordered().into_iter().collect();
-                self.contexts.map.insert(
-                    context.clone(),
-                    PhraseCounter {
-                        map: ordered_context_phrases,
-                    },
-                );
-            }
+            self.merge(&selmr);
         }
         Ok(())
+    }
+
+    /// merge the context of another SELMR struct into the current one
+    pub fn merge(&mut self, other: &SELMR) {
+        // merge the content of s with that of the current struct
+        for (new_p, new_contexts) in other.phrases.map.iter() {
+            for (new_c, new_count) in new_contexts.map.iter() {
+                self.phrases
+                    .map
+                    .entry(new_p.clone())
+                    .or_default()
+                    .map
+                    .entry(new_c.clone())
+                    .and_modify(|count| *count += new_count)
+                    .or_insert(*new_count);
+            }
+        }
+        self.create_and_order_contexts();
+    }
+
+    fn create_and_order_contexts(&mut self) {
+        // the self.contexts is derived from self.phrases, and must be ordered
+        let mut unordered_contexts = HashMap::<Context, Counter<Phrase>>::new();
+        for (p, contexts) in self.phrases.map.iter() {
+            for (c, &n) in contexts.map.iter() {
+                unordered_contexts
+                    .entry(c.clone())
+                    .or_default()
+                    .insert(p.clone(), n);
+            }
+        }
+        self.contexts = ContextMap::new();
+        for (context, context_phrases) in unordered_contexts.iter() {
+            let ordered_context_phrases: IndexMap<Phrase, usize> =
+                context_phrases.most_common_ordered().into_iter().collect();
+            self.contexts.map.insert(
+                context.clone(),
+                PhraseCounter {
+                    map: ordered_context_phrases,
+                },
+            );
+        }
     }
 
     /// add text data to SELMR data structure
@@ -209,17 +241,38 @@ impl SELMR {
     /// # Example
     ///
     /// ```
-    /// let s = selmr::selmr::SELMR::new(1, 3, 1, 3, 1, 1, "en");
+    /// let s = selmr::selmr::SELMR::new();
     /// ```
-    pub fn add<'a>(&mut self, text: &str) {
+    pub fn add<'a>(
+        &mut self,
+        text: &str,
+        min_phrase_len: usize,
+        max_phrase_len: usize,
+        min_left_context_len: usize,
+        max_left_context_len: usize,
+        min_right_context_len: usize,
+        max_right_context_len: usize,
+        min_phrase_keys: usize,
+        min_context_keys: usize,
+        language: &str,
+    ) {
         let binding = text.to_string();
         let text = &tokenizer::tokenize(&binding);
-
+        self.params = Params {
+            min_phrase_len,
+            max_phrase_len,
+            min_left_context_len,
+            max_left_context_len,
+            min_right_context_len,
+            max_right_context_len,
+            min_phrase_keys,
+            min_context_keys,
+            language: language.to_string(),
+        };
         info!(target: "selmr", "adding {} tokens", text.len());
-
         let mut phrases = HashMap::<&'a [&'a str], Counter<(&'a [&'a str], &'a [&'a str])>>::new();
-        for l_len in self.params.min_context_len..self.params.max_context_len + 1 {
-            for r_len in self.params.min_context_len..self.params.max_context_len + 1 {
+        for l_len in self.params.min_left_context_len..self.params.max_left_context_len + 1 {
+            for r_len in self.params.min_right_context_len..self.params.max_right_context_len + 1 {
                 let mut new_phrases =
                     HashMap::<&'a [&'a str], Counter<(&'a [&'a str], &'a [&'a str])>>::new();
                 let mut new_contexts =
@@ -251,7 +304,7 @@ impl SELMR {
                 for (p, ps) in new_phrases.iter() {
                     for ((l, r), count) in ps.iter() {
                         if new_contexts.contains(&(*l, *r)) {
-                            if l.len() > 1 {
+                            if l.len() > self.params.min_left_context_len {
                                 if let Some(cs) = phrases.get(p) {
                                     if let Some(context_count) = cs.get(&(&l[1..], r)) {
                                         if *context_count != *count {
@@ -260,7 +313,7 @@ impl SELMR {
                                     }
                                 }
                             }
-                            if r.len() > 1 {
+                            if r.len() > self.params.min_right_context_len {
                                 if let Some(cs) = phrases.get(p) {
                                     if let Some(context_count) = cs.get(&(l, &r[..r.len() - 1])) {
                                         if *context_count != *count {
@@ -269,7 +322,9 @@ impl SELMR {
                                     }
                                 }
                             }
-                            if l.len() > 1 && r.len() > 1 {
+                            if l.len() > self.params.min_left_context_len
+                                && r.len() > self.params.min_right_context_len
+                            {
                                 if let Some(cs) = phrases.get(p) {
                                     if let Some(context_count) =
                                         cs.get(&(&l[1..], &r[..r.len() - 1]))
@@ -280,7 +335,9 @@ impl SELMR {
                                     }
                                 }
                             }
-                            if l.len() == 1 && r.len() == 1 {
+                            if l.len() == self.params.min_left_context_len
+                                && r.len() == self.params.min_right_context_len
+                            {
                                 phrases.entry(p).or_default()[&(*l, *r)] = *count;
                             }
                         }
@@ -332,229 +389,247 @@ impl SELMR {
 
     /// Returns the most similar phrases based on common contexts
     ///
-    /// # Example
+    /// # Example 1
     ///
     /// ```
-    /// let mut s = selmr::selmr::SELMR::new(1, 3, 1, 3, 1, 1, "en");
-    /// s.add("My sister lives in the city. My brother lives in the mountains.");
-    /// let actual = s.most_similar("sister", None, 25, 25, 15).unwrap();
-    /// let expect = [("brother".to_string(), 1), ("sister".to_string(), 1)]
+    /// use selmr::selmr::SELMR;
+    /// use selmr::text_structs::Phrase;
+    ///
+    /// let mut s = SELMR::new();
+    /// s.add("a 1 b c. a 2 b c. a 2 b d.", 1, 3, 1, 3, 1, 3, 1, 1, "");
+    /// let actual = s.most_similar(Phrase::new("2"), None, 25, 25, 15, "count").unwrap();
+    /// let expect = [(Phrase::new("2"), 5.0), (Phrase::new("1"), 1.0)]
     ///     .iter().cloned().collect::<Vec<_>>();
     /// assert_eq!(actual, expect);
     /// ```
-    #[pyo3(signature = (phrase, context=None, topcontexts=25, topphrases=25, topn=15))]
+    ///
+    /// # Example 2
+    ///
+    /// ```
+    /// use selmr::selmr::SELMR;
+    /// use selmr::text_structs::Phrase;
+    ///
+    /// let mut s = selmr::selmr::SELMR::new();
+    /// s.add("a 1 b c. a 2 b c. a 2 b d.", 1, 3, 1, 3, 1, 3, 1, 1, "");
+    /// let actual = s.most_similar(Phrase::new("2"), None, 25, 25, 15, "jaccard").unwrap();
+    /// let expect = [(Phrase::new("2"), 1.0), (Phrase::new("1"), 0.2)]
+    ///     .iter().cloned().collect::<Vec<_>>();
+    /// assert_eq!(actual, expect);
+    /// ```
+    ///
+    /// # Example 3
+    ///
+    /// ```
+    /// use selmr::selmr::SELMR;
+    /// use selmr::text_structs::Phrase;
+    ///
+    /// let mut s = selmr::selmr::SELMR::new();
+    /// s.add("a 1 b c. a 2 b c. a 2 b d.", 1, 3, 1, 3, 1, 3, 1, 1, "");
+    /// let actual = s.most_similar(Phrase::new("2"), None, 25, 25, 15, "weighted_jaccard").unwrap();
+    /// let expect = [(Phrase::new("2"), 1.0), (Phrase::new("1"), 0.16666667)]
+    ///     .iter().cloned().collect::<Vec<_>>();
+    /// assert_eq!(actual, expect);
+    /// ```
     pub fn most_similar(
         &self,
-        phrase: &str,
-        context: Option<(&str, &str)>,
+        phrase: Phrase,
+        context: Option<Context>,
         topcontexts: usize,
         topphrases: usize,
         topn: usize,
-    ) -> Result<Vec<(String, usize)>, PyErr> {
+        measure: &str,
+    ) -> Result<Vec<(Phrase, f32)>, String> {
+        if let Some(multiset) = self.get_phrase_contexts(&phrase, topcontexts) {
+            self.most_similar_from_multiset(
+                &multiset,
+                context,
+                topcontexts,
+                topphrases,
+                topn,
+                measure,
+            )
+        } else {
+            Err(format!("Phrase not found: {}", phrase))
+        }
+    }
+
+    /// Find the most similar phrases from a given multiset
+    pub fn most_similar_from_multiset(
+        &self,
+        multiset: &HashMap<&Context, &usize>,
+        context: Option<Context>,
+        topcontexts: usize,
+        topphrases: usize,
+        topn: usize,
+        measure: &str,
+    ) -> Result<Vec<(Phrase, f32)>, String> {
+        match self.get_phrases_to_evaluate(&multiset, context, topphrases) {
+            Ok(phrases_to_evaluate) => self.most_similar_from_phrases(
+                multiset,
+                &phrases_to_evaluate,
+                topcontexts,
+                topn,
+                measure,
+            ),
+            Err(e) => Err(e),
+        }
+    }
+
+    /// Find all phrases that fit a multiset of contexts
+    pub fn get_phrases_to_evaluate(
+        &self,
+        input_phrase_multiset: &HashMap<&Context, &usize>,
+        context: Option<Context>,
+        topphrases: usize,
+    ) -> Result<HashSet<&Phrase>, String> {
         // the most_similar function returns the phrases with the most contexts in common with the input phrase
         // first: contexts of the input phrase
-        let cs = self.phrases.map.get(&Phrase {
-            value: String::from(phrase),
-        });
-        // the counter that will contain the results
-        let mut most_similar = Counter::<Phrase>::new();
-        match cs {
-            Some(cs) => {
-                // take the most common contexts of the input phrase
-                let most_common_cs: HashSet<_> = cs.map[..min(cs.map.len(), topcontexts)]
-                    .into_iter()
-                    .map(|x| x.0)
-                    .collect();
-                match context {
-                    // if there is an input context then retrieve phrases that fit that context
-                    // the intersection between the context of the input phrase
-                    // and the contexts of each phrase that fits the input context
-                    Some(context) => {
-                        let c = Context {
-                            left_value: String::from(context.0),
-                            right_value: String::from(context.1),
-                        };
-                        let len_p = min(self.contexts.map[&c].map.len(), topphrases);
-                        let ps: HashSet<_> = self.contexts.map[&c].map[..len_p]
-                            .into_iter()
-                            .map(|x| x.0)
-                            .collect();
-                        for p in &ps {
-                            let len_c = min(self.phrases.map[&p].map.len(), topcontexts);
-                            let p_cs: HashSet<_> = self.phrases.map[&p].map[..len_c]
-                                .into_iter()
-                                .map(|x| x.0)
-                                .collect();
-                            most_similar[&p] =
-                                p_cs.intersection(&most_common_cs).collect::<Vec<_>>().len();
-                        }
-                    }
-                    None => {
-                        // if there is not input context then
-                        // for each context in the most common contexts cs update
-                        // the counter with the phrases that fit in that context
-                        let mut ps: HashSet<_> = HashSet::new();
-                        for c in &most_common_cs {
-                            let len_c = min(self.contexts.map[&c].map.len(), topphrases);
-                            for p in self.contexts.map[&c].map[..len_c]
-                                .into_iter()
-                                .map(|x| x.0)
-                                .collect::<Vec<_>>()
-                            {
-                                ps.insert(p);
-                            }
-                        }
-                        for p in ps {
-                            let len_p = min(self.phrases.map[&p].map.len(), topcontexts);
-                            let p_cs: HashSet<_> = self.phrases.map[&p].map[..len_p]
-                                .into_iter()
-                                .map(|x| x.0)
-                                .collect();
-                            most_similar[&p] =
-                                most_common_cs.intersection(&p_cs).collect::<Vec<_>>().len();
-                        }
-                    }
-                }
-                // the count of occurrences is not used because then
-                // phrases with high occurrences are preferred in the output
-            }
-            None => {
-                return Err(PyErr::new::<PyTypeError, _>(
-                    "Phrase was not found when calling most_similar",
-                ))
-            }
-        }
-        Ok(most_similar
-            .k_most_common_ordered(topn)
-            .iter()
-            .map(|(phrase, count)| (phrase.to_string(), *count))
-            .collect())
-    }
-
-    /// Returns the most common contexts of a phrase
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// let mut s = selmr::selmr::SELMR::new(1, 3, 1, 3, 1, 1, "en");
-    /// s.add("My sister lives in the city. My brother lives in the mountains.");
-    /// let actual = s.get_contexts(Some("sister"), None, 15).unwrap();
-    /// let expect = [("My ... lives".to_string(), 1)]
-    ///     .iter().cloned().collect::<Vec<_>>();
-    /// assert_eq!(actual, expect);
-    /// ```
-    #[pyo3(signature = (phrase=None, phrases=None, topn=15))]
-    pub fn get_contexts(
-        &self,
-        phrase: Option<&str>,
-        phrases: Option<Vec<&str>>,
-        topn: usize,
-    ) -> Result<Vec<(String, usize)>, PyErr> {
-        match phrase {
-            Some(phrase) => {
-                let contexts = self.phrases.map.get(&Phrase {
-                    value: String::from(phrase),
-                });
-                match contexts {
-                    Some(contexts) => Ok(contexts.map[..min(topn, contexts.map.len())]
-                        .iter()
-                        .map(|(context, count)| (context.to_string(), *count))
-                        .collect()),
-                    None => Err(PyErr::new::<PyTypeError, _>(
-                        "Contexts of phrase not found: ".to_owned() + phrase,
-                    )),
-                }
-            }
-            None => match phrases {
-                Some(_phrases) => {
-                    let mut new_counter = Counter::<Context>::new();
-                    for phrase in _phrases {
-                        let contexts = self.phrases.map.get(&Phrase {
-                            value: String::from(phrase),
-                        });
-                        match contexts {
-                            Some(contexts) => {
-                                let len_c = min(contexts.map.len(), topn);
-                                let contexts = &contexts.map[..len_c];
-                                for (context, count) in contexts.iter() {
-                                    new_counter
-                                        .entry(context.clone())
-                                        .and_modify(|c| *c += count)
-                                        .or_insert(*count);
-                                }
-                            }
-                            None => {
-                                return Err(PyErr::new::<PyTypeError, _>(
-                                    "Contexts of phrase not found: ".to_owned() + phrase,
-                                ))
-                            }
-                        }
-                    }
-                    let ordered_new_counter: IndexMap<Context, usize> =
-                        new_counter.most_common_ordered().into_iter().collect();
-                    Ok(ordered_new_counter[..min(new_counter.len(), topn)]
-                        .iter()
-                        .map(|(context, count)| (context.to_string(), *count))
-                        .collect())
-                }
-                None => Err(PyErr::new::<PyTypeError, _>(
-                    "No phrase(s) given when calling get_contexts",
-                )),
-            },
-        }
-    }
-
-    /// Returns the most common phrases of a context
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// let mut s = selmr::selmr::SELMR::new(1, 3, 1, 3, 1, 1, "en");
-    /// s.add("My sister lives in the city. My brother lives in the mountains.");
-    /// let actual = s.get_phrases(Some(("My", "lives")), 15).unwrap();
-    /// let expect = [("brother".to_string(), 1), ("sister".to_string(), 1)]
-    ///     .iter().cloned().collect::<Vec<_>>();
-    /// assert_eq!(actual, expect);
-    /// ```
-    #[pyo3(signature = (context=None, topn=15))]
-    pub fn get_phrases(
-        &self,
-        context: Option<(&str, &str)>,
-        topn: usize,
-    ) -> Result<Vec<(String, usize)>, PyErr> {
+        let mut phrases_to_evaluate: HashSet<&Phrase> = HashSet::new();
         match context {
             Some(context) => {
-                let phrases = self.contexts.map.get(&Context {
-                    left_value: String::from(context.0),
-                    right_value: String::from(context.1),
-                });
-                match phrases {
-                    Some(phrases) => Ok(phrases.map[..min(phrases.map.len(), topn)]
-                        .iter()
-                        .map(|(phrase, count)| (phrase.to_string(), *count))
-                        .collect()),
-                    None => Err(PyErr::new::<PyTypeError, _>(
-                        "Context not found when calling get_phrases",
-                    )),
+                // if there is an input context then
+                // retrieve phrases that fit that context
+                if let Some(items) = self.get_context_phrases(&context, topphrases) {
+                    phrases_to_evaluate.extend(items.keys());
                 }
-            },
-            None => Err(PyErr::new::<PyTypeError, _>(
-                "No context given when calling get_phrases",
-            )),
+            }
+            None => {
+                // if there is no input context then
+                // for each context in the most common contexts cs update
+                for c in input_phrase_multiset.keys() {
+                    if let Some(items) = self.get_context_phrases(c, topphrases) {
+                        phrases_to_evaluate.extend(items.keys());
+                    }
+                }
+            }
         }
+        Ok(phrases_to_evaluate)
+    }
+
+    /// Find most similar phrases from a list of phrases to evaluate
+    pub fn most_similar_from_phrases(
+        &self,
+        input_phrase_multiset: &HashMap<&Context, &usize>,
+        phrases_to_evaluate: &HashSet<&Phrase>,
+        topcontexts: usize,
+        topn: usize,
+        measure: &str,
+    ) -> Result<Vec<(Phrase, f32)>, String> {
+        let mut result = Vec::<(Phrase, f32)>::new();
+        match measure {
+            "count" | "jaccard" => {
+                let set_b: HashSet<&Context> = input_phrase_multiset.keys().cloned().collect();
+                for phrase in phrases_to_evaluate {
+                    let set_a = &self
+                        .get_phrase_contexts(phrase, topcontexts)
+                        .unwrap()
+                        .keys()
+                        .cloned()
+                        .collect();
+                    let value: f32 = match measure {
+                        "count" => count_index(set_a, &set_b),
+                        "jaccard" => jaccard_index(set_a, &set_b),
+                        _ => 0.0,
+                    };
+                    result.push((phrase.clone().clone(), value));
+                }
+            }
+            "weighted_jaccard" => {
+                // iterate over first n phrases that fit context c
+                for phrase in phrases_to_evaluate {
+                    if let Some(multiset_a) = &self.phrases.get(phrase, topcontexts) {
+                        let value: f32 = weighted_jaccard_index(multiset_a, input_phrase_multiset);
+                        result.push((phrase.clone().clone(), value));
+                    } else {
+                        return Err(format!("Phrase not found: {}", phrase))
+                    }
+                }
+            }
+            m => return Err(format!("Measure not implemented: {}", m)),
+        }
+        println!("{:?}", result);
+        // sort result based on measure
+        result.sort_by(|a, b| b.1.partial_cmp(&a.1).unwrap());
+        // select topn results
+        result = result[..min(result.len(), topn)].to_vec();
+        Ok(result)
+    }
+
+    /// Returns the context multiset of a phrase
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use std::collections::HashMap;
+    /// use selmr::selmr::SELMR;
+    /// use selmr::text_structs::{Phrase, Context};
+    ///
+    /// let mut s = selmr::selmr::SELMR::new();
+    /// s.add("a 1 b c. a 2 b c. a 2 b d.", 1, 3, 1, 3, 1, 3, 1, 1, "");
+    /// let actual = s.get_phrase_contexts(&Phrase::new("1"), 15).unwrap();
+    /// let expect = [(&Context::new("a ... b"), &1)].iter().cloned().collect::<HashMap<_, _>>();
+    /// assert_eq!(actual, expect);
+    /// ```
+    ///
+    pub fn get_phrase_contexts(
+        &self,
+        phrase: &Phrase,
+        n: usize,
+    ) -> Option<HashMap<&Context, &usize>> {
+        self.phrases.get(phrase, n)
+    }
+
+    /// Returns the context multiset of a list of phrases
+    pub fn get_phrases_contexts(
+        &self,
+        phrases: &Vec<Phrase>,
+        n: usize,
+    ) -> Result<HashMap<Context, usize>, String> {
+        let mut result = HashMap::<Context, usize>::new();
+        for phrase in phrases {
+            if let Some(ps) = self.phrases.get(phrase, n) {
+                for (&p, &count) in &ps {
+                    *result.entry(p.clone()).or_insert(0) += count;
+                }
+            } else {
+                return Err(format!("Phrase not found: {}", phrase));
+            }
+        }
+        Ok(result)
+    }
+
+    /// Returns the phrase multiset of a context
+    pub fn get_context_phrases(
+        &self,
+        context: &Context,
+        n: usize,
+    ) -> Option<HashMap<&Phrase, &usize>> {
+        self.contexts.get(context, n)
+    }
+
+    /// Returns the phrase multiset of a list of contexts
+    pub fn get_contexts_phrases(
+        &self,
+        contexts: &Vec<Context>,
+        n: usize,
+    ) -> Result<HashMap<Phrase, usize>, String> {
+        let mut result = HashMap::<Phrase, usize>::new();
+        for context in contexts {
+            if let Some(cs) = self.contexts.get(context, n) {
+                for (&c, &count) in &cs {
+                    *result.entry(c.clone()).or_insert(0) += count;
+                }
+            } else {
+                return Err(format!("Context not found: {}", context));
+            }
+        }
+        Ok(result)
     }
 
     /// Prunes a SELMR data structure
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// ```
-    #[pyo3(signature = (topn_phrases=50, topn_contexts=50))]
     pub fn prune(&mut self, topn_phrases: usize, topn_contexts: usize) {
         let mut new_phrases = PhraseMap::new();
         let mut new_contexts = ContextMap::new();
-
         for p in self.phrases.map.keys() {
             let mut new_counter = ContextCounter {
                 map: IndexMap::<Context, usize>::new(),
@@ -580,11 +655,6 @@ impl SELMR {
     }
 
     /// Filters a SELMR data structure on given regexes
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// ```
     pub fn matches(
         &self,
         tuple: (&str, &str, &str),